2.13. Рефакторинг. Функция createOffer. WIP + сamelCase
