+++ mocks/mock-data.tsv
@@ -1,2 +1,2 @@
 Beautiful apartment in Paris	Spacious and bright apartment in the heart of Paris with Eiffel Tower view. Perfect for romantic getaway or business trip.	2024-01-15	Paris	preview-paris.jpg	image1.jpg;image2.jpg;image3.jpg;image4.jpg;image5.jpg;image6.jpg	true	false	4.8	apartment	2	4	15000	Breakfast;Air conditioning;Laptop friendly workspace;Fridge	John Doe	john@example.com	john-avatar.jpg	password123	pro	0	48.85661	2.351499
-Cozy house in Amsterdam	Charming traditional Dutch house with canal view. Located in quiet area close to city center.	2024-01-20	Amsterdam	preview-amsterdam.jpg	house1.jpg;house2.jpg;house3.jpg;house4.jpg;house5.jpg;house6.jpg	false	true	4.5	house	3	6	25000	Breakfast;Washer;Towels;Fridge;Baby seat	Alice Smith	alice@example.com		securepass	обычный	0	52.370216	4.895168
+Cozy house in Amsterdam	Charming traditional Dutch house with canal view. Located in quiet area close to city center.	2024-01-20	Amsterdam	preview-amsterdam.jpg	house1.jpg;house2.jpg;house3.jpg;house4.jpg;house5.jpg;house6.jpg	false	true	4.5	house	3	6	25000	Breakfast;Washer;Towels;Fridge;Baby seat	Alice Smith	alice@example.com		securepass	regular	0	52.370216	4.895168
