+++ src/shared/helpers/index.ts
@@ -1,1 +1,6 @@
 export * from './common.js';
+export * from './../types/cities-name.enum.js';
+export * from './../types/conveniences.enum.js';
+export * from './../types/house.type.enum.js';
+export * from './../types/user.type.js';
+export * from './../types/offer.type.js';
